Improve reliability of network unavailable dialog
Use a large timeout.

Speculation:

We're in a happy place here.  The main issue at this point is that DISCONNECT followed by a PLAY prior to the window popping up -- will cause the system to hang.... but only once the OPPONENT has returned a stone!  The play itself goes through to the opponent, which is good.

Can we prevent player input during this time? (By aggressively polling state but not causing a strobe light?)

Can we make sure we don't end up hung during the "return serve?"

These are future issues.